INS-447: transport copy-paste code refactored
